circular dependency fix
